refactor : Quantity는 공통영역이므로 common으로 이동
