+++ src/main/java/kitchenpos/menu/dto/MenuProductResponse.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.dto;
 
 import kitchenpos.menu.domain.MenuProduct;
-import kitchenpos.menu.domain.Quantity;
+import kitchenpos.common.domain.Quantity;
 
 public class MenuProductResponse {
     private Long seq;
